+++ models.py
@@ -1,7 +1,9 @@
 import datetime
 import random
 import re
+import subprocess
 
+import praw
 from tinydb import TinyDB, Query
 
 import bot_logger
@@ -9,6 +11,16 @@ import config
 import crypto
 import user_function
 import utils
+
+B58_DIGITS = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz'
+
+
+def base58_is_valid(string):
+    for car in string:
+        if car not in B58_DIGITS:
+            return False
+
+    return True
 
 
 class Tip(object):
@@ -115,23 +127,11 @@ class Tip(object):
         # import user
         self.receiver = User(arr_tip['receiver'])
         self.sender = User(arr_tip['sender'])
+        del arr_tip['receiver']
+        del arr_tip['sender']
 
-        self.id = arr_tip['id']
-        self.amount = arr_tip['amount']
-
-        if 'message_fullname' in arr_tip.keys():
-            self.message_fullname = arr_tip['message_fullname']
-
-        if 'finish' in arr_tip.keys():
-            self.finish = arr_tip['finish']
-
-        if 'tx_id' in arr_tip.keys():
-            self.tx_id = arr_tip['tx_id']
-
-        if 'status' in arr_tip.keys():
-            self.status = arr_tip['status']
-
-        self.time = arr_tip['time']
+        for key in arr_tip.keys():
+            setattr(self, key, arr_tip[key])
 
         return self
 
@@ -215,6 +215,11 @@ class User(object):
         # register in users table
         UserStorage.add_address(self.username, self.address)
 
+    def send_private_message(self, title, content):
+        if self.reddit is None:
+            self.reddit = praw.Reddit(config.bot_name)
+        self.reddit.redditor(self.username).message(title, content)
+
 
 class UserStorage:
     @staticmethod
@@ -227,7 +232,10 @@ class UserStorage:
         data = table.count(user_db.address == address)
 
         if data == 0:
-            table.insert({"type": "simple", "address": address, "coin": "doge", "enable": active})
+            table.insert({"type": "simple", "address": address, "coin": "doge", "enable": False})
+
+            if active is True:
+                UserStorage.active_user_address(username, address)
         else:
             bot_logger.logger.error("address %s already registered for  %s " % (str(address), str(username)))
 
@@ -271,3 +279,117 @@ class UserStorage:
             return data[0].get('address')
         else:
             bot_logger.logger.error("get address of un-registered user  %s " % (str(username)))
+
+    @classmethod
+    def active_user_address(cls, username, address):
+        if UserStorage.exist(username):
+            db = TinyDB(config.user_file)
+            table = db.table(username)
+
+            # check if address not already exist
+            user_db = Query()
+            data = table.count(user_db.address == address)
+
+            if data == 0:
+                # disable all other address
+                table.update(set("enable", False))
+
+                # enable only one
+                table.update(set("enable", True), user_db.address == address)
+            else:
+                bot_logger.logger.error("active a not found address (%s)  of user  %s " % (str(address), str(username)))
+
+        else:
+            bot_logger.logger.error("active address of un-registered user  %s " % (str(username)))
+
+
+class VanityGenRequest(object):
+    """Class to represent an user"""
+
+    def __init__(self, user):
+        self.user = User(user)
+        self.pattern = None
+        self.difficulty = None
+        self.address = None
+        self.private_key = None
+        self.duration = 0
+
+        self.id = random.randint(0, 99999999)
+
+        # to check if we move funds of user
+        self.use = None
+
+    def parse_message(self, message_to_parse, rpc=None):
+        split_message = message_to_parse.split()
+
+        # parse message like : +vanity use Dpatern
+        donate_index = split_message.index('+vanity')
+        use_address = split_message[donate_index + 1]
+        pattern = split_message[donate_index + 2]
+
+        if use_address == "use":
+            self.use = True
+        if use_address == "not-use":
+            self.use = False
+
+        # check patern is ok
+        if pattern[0] == "D" and base58_is_valid(pattern):
+            self.pattern = pattern
+
+    def save_resquest(self):
+        if self.pattern is not None:
+            db = TinyDB(config.vanitygen)
+            db.insert({
+                "id": self.id,
+                "user": self.user.username,
+                "use": self.use,
+                "pattern": self.pattern,
+                "finish": False,
+                "address": self.address,
+                "difficulty": self.difficulty,
+                "duration": 0
+            })
+            return True
+        else:
+            return False
+
+    def create_from_array(self, arr_vanity):
+        self.user = User(arr_vanity['user'])
+        del arr_vanity['user']
+
+        for key in arr_vanity.keys():
+            setattr(self, key, arr_vanity[key])
+
+        return self
+
+    def generate(self):
+        out = subprocess.check_output(['vanitygen', '-X', '30', str(self.pattern)], stderr=subprocess.STDOUT)
+        line = out.split('\n')
+        self.difficulty = str((line[0]).split(':')[1]).strip()
+        self.address = str((line[1]).split(':')[1]).strip()
+        self.private_key = str((line[2]).split(':')[1]).strip()
+
+    def move_funds(self, tx_queue, failover_time):
+        if self.use is True:
+            amount = self.user.get_balance()
+            if crypto.tip_user(self.user.address, self.address, amount, tx_queue, failover_time):
+                # update user storage file
+                UserStorage.add_address(self.user.username, self.address)
+
+                # Todo : update history of user
+
+    def import_address(self):
+        rpc = crypto.get_rpc()
+        rpc.importprivkey(self.private_key, "reddit-vanity-" + self.user.username, false)
+
+        # on import success clean key from memory
+        self.private_key = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
+        self.private_key = None
+
+    def update_data(self):
+        db = TinyDB(config.vanitygen)
+        vanity_db = Query()
+
+        db.update(set("finish", True), vanity_db.id == self.id)
+        db.update(set("difficulty", self.difficulty), vanity_db.id == self.id)
+        db.update(set("duration", self.duration), vanity_db.id == self.id)